Remove use of Junit4, clean up and consolidate tests in tm module
