+++ services-custom/s3-transfer-manager/src/it/java/software/amazon/awssdk/transfer/s3/S3IntegrationTestBase.java
@@ -24,6 +24,7 @@ import software.amazon.awssdk.services.s3.S3AsyncClient;
 import software.amazon.awssdk.services.s3.S3AsyncClientBuilder;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.S3ClientBuilder;
+import software.amazon.awssdk.services.s3.internal.crt.S3CrtAsyncClient;
 import software.amazon.awssdk.services.s3.model.BucketLocationConstraint;
 import software.amazon.awssdk.services.s3.model.CreateBucketConfiguration;
 import software.amazon.awssdk.services.s3.model.CreateBucketRequest;
@@ -52,6 +53,10 @@ public class S3IntegrationTestBase extends AwsTestBase {
 
     protected static S3AsyncClient s3Async;
 
+    protected static S3AsyncClient s3CrtAsync;
+
+    protected static S3TransferManager tm;
+
     /**
      * Loads the AWS account info for the integration tests and creates an S3
      * client for tests to use.
@@ -62,12 +67,21 @@ public class S3IntegrationTestBase extends AwsTestBase {
         System.setProperty("aws.crt.debugnative", "true");
         s3 = s3ClientBuilder().build();
         s3Async = s3AsyncClientBuilder().build();
+        s3CrtAsync = S3CrtAsyncClient.builder()
+                                     .credentialsProvider(CREDENTIALS_PROVIDER_CHAIN)
+                                     .region(DEFAULT_REGION)
+                                     .build();
+        tm = S3TransferManager.builder()
+                              .s3AsyncClient(s3CrtAsync)
+                              .build();
     }
 
     @AfterAll
     public static void cleanUp() {
         s3.close();
         s3Async.close();
+        s3CrtAsync.close();
+        tm.close();
         CrtResource.waitForNoResources();
     }
