+++ services-custom/s3-transfer-manager/src/it/java/software/amazon/awssdk/transfer/s3/S3TransferManagerUploadDirectoryIntegrationTest.java
@@ -19,19 +19,30 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static software.amazon.awssdk.testutils.service.S3BucketUtils.temporaryBucketName;
 import static software.amazon.awssdk.utils.IoUtils.closeQuietly;
 
+import com.google.common.jimfs.Configuration;
+import com.google.common.jimfs.Jimfs;
 import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.FileSystem;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
+import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
+import org.apache.commons.codec.binary.Hex;
 import org.apache.commons.lang3.RandomStringUtils;
+import org.assertj.core.util.Sets;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import software.amazon.awssdk.core.sync.ResponseTransformer;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
 import software.amazon.awssdk.services.s3.S3Client;
@@ -46,9 +57,7 @@ public class S3TransferManagerUploadDirectoryIntegrationTest extends S3Integrati
     private static final Logger log = Logger.loggerFor(S3TransferManagerUploadDirectoryIntegrationTest.class);
     private static final String TEST_BUCKET = temporaryBucketName(S3TransferManagerUploadDirectoryIntegrationTest.class);
 
-    private static S3TransferManager tm;
     private static Path directory;
-    private static S3Client s3Client;
     private static String randomString;
 
     @BeforeAll
@@ -57,16 +66,6 @@ public class S3TransferManagerUploadDirectoryIntegrationTest extends S3Integrati
         createBucket(TEST_BUCKET);
         randomString = RandomStringUtils.random(100);
         directory = createLocalTestDirectory();
-
-        tm = S3TransferManager.builder()
-                              .s3AsyncClient(S3AsyncClient.builder()
-                                                          .credentialsProvider(CREDENTIALS_PROVIDER_CHAIN)
-                                                          .region(DEFAULT_REGION)
-                                                          .build()).build();
-
-        s3Client = S3Client.builder()
-                           .credentialsProvider(CREDENTIALS_PROVIDER_CHAIN).region(DEFAULT_REGION)
-                           .build();
     }
 
     @AfterAll
@@ -82,9 +81,6 @@ public class S3TransferManagerUploadDirectoryIntegrationTest extends S3Integrati
         } catch (Exception exception) {
             log.warn(() -> "Failed to delete s3 bucket " + TEST_BUCKET, exception);
         }
-
-        closeQuietly(tm, log.logger());
-        closeQuietly(s3Client, log.logger());
         S3IntegrationTestBase.cleanUp();
     }
 
@@ -98,7 +94,7 @@ public class S3TransferManagerUploadDirectoryIntegrationTest extends S3Integrati
         assertThat(completedDirectoryUpload.failedTransfers()).isEmpty();
 
         List<String> keys =
-            s3Client.listObjectsV2Paginator(b -> b.bucket(TEST_BUCKET).prefix(prefix)).contents().stream().map(S3Object::key)
+            s3.listObjectsV2Paginator(b -> b.bucket(TEST_BUCKET).prefix(prefix)).contents().stream().map(S3Object::key)
                     .collect(Collectors.toList());
 
         assertThat(keys).containsOnly(prefix + "/bar.txt", prefix + "/foo/1.txt", prefix + "/foo/2.txt");
@@ -129,7 +125,7 @@ public class S3TransferManagerUploadDirectoryIntegrationTest extends S3Integrati
         assertThat(completedDirectoryUpload.failedTransfers()).isEmpty();
 
         List<String> keys =
-            s3Client.listObjectsV2Paginator(b -> b.bucket(TEST_BUCKET).prefix(prefix)).contents().stream().map(S3Object::key)
+            s3.listObjectsV2Paginator(b -> b.bucket(TEST_BUCKET).prefix(prefix)).contents().stream().map(S3Object::key)
                     .collect(Collectors.toList());
 
         assertThat(keys).containsOnly(prefix + "0bar.txt", prefix + "0foo01.txt", prefix + "0foo02.txt");
@@ -153,9 +149,60 @@ public class S3TransferManagerUploadDirectoryIntegrationTest extends S3Integrati
               .get(10, TimeUnit.SECONDS);
         assertThat(result.failedTransfers()).isEmpty();
 
-        s3Client.listObjectsV2Paginator(b -> b.bucket(TEST_BUCKET).prefix(prefix)).contents().forEach(object -> {
+        s3.listObjectsV2Paginator(b -> b.bucket(TEST_BUCKET).prefix(prefix)).contents().forEach(object -> {
             verifyContent(object.key(), "bar.txt" + randomString);
         });
+    }
+
+    public static Collection<String> prefix() {
+        return Arrays.asList(
+            /* ASCII, 1-byte UTF-8 */
+            "E",
+            /* ASCII, 2-byte UTF-8 */
+            "É",
+            /* Non-ASCII, 2-byte UTF-8 */
+            "Ũ",
+            /* Non-ASCII, 3-byte UTF-8 */
+            "स",
+            /* Non-ASCII, 4-byte UTF-8 */
+            "\uD808\uDC8C"
+        );
+    }
+
+    /**
+     * Tests the behavior of traversing local directories with special Unicode characters in their path name. These characters have
+     * known to be problematic when using Java's old File API or with Windows (which uses UTF-16 for file-name encoding).
+     */
+    @ParameterizedTest
+    @MethodSource("prefix")
+    void uploadDirectory_fileNameWithUnicode_traversedCorrectly(String directoryPrefix) throws IOException {
+        Path testDirectory = null;
+        try {
+            log.info(() -> "Testing directory prefix: " + directoryPrefix);
+            log.info(() -> "UTF-8 bytes: " + Hex.encodeHexString(directoryPrefix.getBytes(StandardCharsets.UTF_8)));
+            log.info(() -> "UTF-16 bytes: " + Hex.encodeHexString(directoryPrefix.getBytes(StandardCharsets.UTF_16)));
+            testDirectory = createLocalTestDirectory(directoryPrefix);
+
+            Path finalTestDirectory = testDirectory;
+            DirectoryUpload uploadDirectory = tm.uploadDirectory(u -> u.sourceDirectory(finalTestDirectory)
+                                                                       .bucket(TEST_BUCKET));
+            CompletedDirectoryUpload completedDirectoryUpload = uploadDirectory.completionFuture().join();
+            assertThat(completedDirectoryUpload.failedTransfers()).isEmpty();
+
+            List<String> keys = s3.listObjectsV2Paginator(b -> b.bucket(TEST_BUCKET))
+                                  .contents()
+                                  .stream()
+                                  .map(S3Object::key)
+                                  .collect(Collectors.toList());
+
+            assertThat(keys).containsOnly("bar.txt", "foo/1.txt", "foo/2.txt");
+
+            keys.forEach(k -> verifyContent(finalTestDirectory, k));
+
+        } finally {
+            FileUtils.cleanUpTestDirectory(testDirectory);
+        }
+
     }
 
     private static Path createLocalTestDirectory() throws IOException {
@@ -171,10 +218,35 @@ public class S3TransferManagerUploadDirectoryIntegrationTest extends S3Integrati
         return directory;
     }
 
+    private Path createLocalTestDirectory(String directoryPrefix) throws IOException {
+        Path dir = Files.createTempDirectory(directoryPrefix);
+        Files.createDirectory(Paths.get(dir + "/foo"));
+        Files.write(dir.resolve("bar.txt"), randomBytes(1024));
+        Files.write(dir.resolve("foo/1.txt"), randomBytes(1024));
+        Files.write(dir.resolve("foo/2.txt"), randomBytes(1024));
+        return dir;
+    }
+
     private static void verifyContent(String key, String expectedContent) {
         String actualContent = s3.getObject(r -> r.bucket(TEST_BUCKET).key(key),
                                       ResponseTransformer.toBytes()).asUtf8String();
 
         assertThat(actualContent).isEqualTo(expectedContent);
+    }
+
+    private void verifyContent(Path testDirectory, String key) {
+        try {
+            byte[] expectedContent = Files.readAllBytes(testDirectory.resolve(key));
+            byte[] actualContent = s3.getObject(r -> r.bucket(TEST_BUCKET).key(key), ResponseTransformer.toBytes()).asByteArray();
+            assertThat(actualContent).isEqualTo(expectedContent);
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
+    private static byte[] randomBytes(long size) {
+        byte[] bytes = new byte[Math.toIntExact(size)];
+        ThreadLocalRandom.current().nextBytes(bytes);
+        return bytes;
     }
 }
