+++ services-custom/s3-transfer-manager/src/test/java/software/amazon/awssdk/transfer/s3/internal/DownloadDirectoryHelperTest.java
@@ -24,24 +24,29 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static software.amazon.awssdk.transfer.s3.util.S3ApiCallMockUtils.stubSuccessfulListObjects;
 
+import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystem;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
+import java.util.stream.Collectors;
+import org.assertj.core.util.Sets;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 import software.amazon.awssdk.core.exception.SdkClientException;
@@ -62,34 +67,33 @@ import software.amazon.awssdk.transfer.s3.progress.LoggingTransferListener;
 import software.amazon.awssdk.transfer.s3.progress.TransferListener;
 
 public class DownloadDirectoryHelperTest {
-    private static FileSystem jimfs;
-    private static Path directory;
+    private static final String DIRECTORY_NAME = "test";
+    private FileSystem fs;
+    private Path directory;
     private Function<DownloadFileRequest, FileDownload> singleDownloadFunction;
     private DownloadDirectoryHelper downloadDirectoryHelper;
     private ListObjectsHelper listObjectsHelper;
 
-    @BeforeAll
-    public static void setup() {
-        jimfs = Jimfs.newFileSystem();
-        directory = jimfs.getPath("test");
-    }
-
-    @AfterAll
-    public static void tearDown() {
-        try {
-            jimfs.close();
-        } catch (IOException e) {
-            // no-op
-        }
-    }
-
     @BeforeEach
     public void methodSetup() {
+        fs = Jimfs.newFileSystem();
+        directory = fs.getPath("test");
         listObjectsHelper = mock(ListObjectsHelper.class);
         singleDownloadFunction = mock(Function.class);
         downloadDirectoryHelper = new DownloadDirectoryHelper(TransferManagerConfiguration.builder().build(),
                                                               listObjectsHelper,
                                                               singleDownloadFunction);
+    }
+
+    @AfterEach
+    public void methodCleanup() throws IOException {
+        fs.close();
+    }
+
+    public static Collection<FileSystem> fileSystems() {
+        return Sets.newHashSet(Arrays.asList(Jimfs.newFileSystem(Configuration.unix()),
+                                             Jimfs.newFileSystem(Configuration.osX()),
+                                             Jimfs.newFileSystem(Configuration.windows())));
     }
 
     @Test
@@ -285,6 +289,142 @@ public class DownloadDirectoryHelperTest {
             assertThat(l.encodingType()).isEqualTo(newEncodingType);
             assertThat(l.maxKeys()).isEqualTo(newMaxKeys);
         });
+    }
+
+
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void downloadDirectory_shouldRecursivelyDownload(FileSystem jimfs) {
+        directory = jimfs.getPath("test");
+        String[] keys = {"1.png", "2020/1.png", "2021/1.png", "2022/1.png", "2023/1/1.png"};
+        stubSuccessfulListObjects(listObjectsHelper, keys);
+        ArgumentCaptor<DownloadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(DownloadFileRequest.class);
+
+        when(singleDownloadFunction.apply(requestArgumentCaptor.capture()))
+            .thenReturn(completedDownload());
+        DirectoryDownload downloadDirectory =
+            downloadDirectoryHelper.downloadDirectory(DownloadDirectoryRequest.builder()
+                                                                              .destination(directory)
+                                                                              .bucket("bucket")
+                                                                              .build());
+        CompletedDirectoryDownload completedDirectoryDownload = downloadDirectory.completionFuture().join();
+        assertThat(completedDirectoryDownload.failedTransfers()).isEmpty();
+
+        List<DownloadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+        actualRequests.forEach(r -> assertThat(r.getObjectRequest().bucket()).isEqualTo("bucket"));
+
+        assertThat(actualRequests.size()).isEqualTo(keys.length);
+
+        verifyDestinationPathForSingleDownload(jimfs, "/", keys, actualRequests);
+    }
+
+    /**
+     * The S3 bucket has the following keys:
+     * abc/def/image.jpg
+     * abc/def/title.jpg
+     * abc/def/ghi/xyz.txt
+     *
+     * if the prefix is "abc/def/", the structure should like this:
+     * image.jpg
+     * title.jpg
+     * ghi
+     *  - xyz.txt
+     */
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void downloadDirectory_withPrefix_shouldStripPrefixInDestinationPath(FileSystem jimfs) {
+        directory = jimfs.getPath("test");
+        String[] keys = {"abc/def/image.jpg", "abc/def/title.jpg", "abc/def/ghi/xyz.txt"};
+        stubSuccessfulListObjects(listObjectsHelper, keys);
+        ArgumentCaptor<DownloadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(DownloadFileRequest.class);
+
+        when(singleDownloadFunction.apply(requestArgumentCaptor.capture()))
+            .thenReturn(completedDownload());
+        DirectoryDownload downloadDirectory =
+            downloadDirectoryHelper.downloadDirectory(DownloadDirectoryRequest.builder()
+                                                                              .destination(directory)
+                                                                              .bucket("bucket")
+                                                                              .listObjectsV2RequestTransformer(l -> l.prefix(
+                                                                                  "abc/def/"))
+                                                                              .build());
+        CompletedDirectoryDownload completedDirectoryDownload = downloadDirectory.completionFuture().join();
+        assertThat(completedDirectoryDownload.failedTransfers()).isEmpty();
+
+        List<DownloadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+
+        assertThat(actualRequests.size()).isEqualTo(keys.length);
+
+        List<String> destinations =
+            actualRequests.stream().map(u -> u.destination().toString())
+                          .collect(Collectors.toList());
+
+        String jimfsSeparator = jimfs.getSeparator();
+
+        List<String> expectedPaths =
+            Arrays.asList("image.jpg", "title.jpg", "ghi/xyz.txt").stream()
+                  .map(k -> DIRECTORY_NAME + jimfsSeparator + k.replace("/",jimfsSeparator)).collect(Collectors.toList());
+        assertThat(destinations).isEqualTo(expectedPaths);
+    }
+
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void downloadDirectory_withDelimiter_shouldHonor(FileSystem jimfs) {
+        directory = jimfs.getPath("test");
+        String delimiter = "|";
+        String[] keys = {"1.png", "2020|1.png", "2021|1.png", "2022|1.png", "2023|1|1.png"};
+        stubSuccessfulListObjects(listObjectsHelper, keys);
+        ArgumentCaptor<DownloadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(DownloadFileRequest.class);
+
+        when(singleDownloadFunction.apply(requestArgumentCaptor.capture())).thenReturn(completedDownload());
+        DirectoryDownload downloadDirectory =
+            downloadDirectoryHelper.downloadDirectory(DownloadDirectoryRequest.builder()
+                                                                              .destination(directory)
+                                                                              .bucket("bucket")
+                                                                              .listObjectsV2RequestTransformer(r -> r.delimiter(delimiter))
+                                                                              .build());
+        CompletedDirectoryDownload completedDirectoryDownload = downloadDirectory.completionFuture().join();
+        assertThat(completedDirectoryDownload.failedTransfers()).isEmpty();
+
+        List<DownloadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+        actualRequests.forEach(r -> assertThat(r.getObjectRequest().bucket()).isEqualTo("bucket"));
+        assertThat(actualRequests.size()).isEqualTo(keys.length);
+
+        verifyDestinationPathForSingleDownload(jimfs, delimiter, keys, actualRequests);
+    }
+
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void downloadDirectory_notDirectory_shouldCompleteFutureExceptionally(FileSystem jimfs) throws IOException {
+        directory = jimfs.getPath("test");
+        Path file = jimfs.getPath("afile" + UUID.randomUUID());
+        Files.write(file, "hellowrold".getBytes(StandardCharsets.UTF_8));
+        assertThatThrownBy(() -> downloadDirectoryHelper.downloadDirectory(DownloadDirectoryRequest.builder().destination(file)
+                                                                                                   .bucket("bucketName").build()).completionFuture().join())
+            .hasMessageContaining("is not a directory").hasCauseInstanceOf(IllegalArgumentException.class);
+    }
+
+    private static DefaultFileDownload completedDownload() {
+        return new DefaultFileDownload(CompletableFuture.completedFuture(CompletedFileDownload.builder()
+                                                                                              .response(GetObjectResponse.builder().build())
+                                                                                              .build()),
+                                       new DefaultTransferProgress(DefaultTransferProgressSnapshot.builder()
+                                                                                                  .transferredBytes(0L)
+                                                                                                  .build()),
+                                       () -> DownloadFileRequest.builder().getObjectRequest(GetObjectRequest.builder().build())
+                                                                .destination(Paths.get("."))
+                                                                .build(),
+                                       null);
+    }
+
+    private static void verifyDestinationPathForSingleDownload(FileSystem jimfs, String delimiter, String[] keys,
+                                                               List<DownloadFileRequest> actualRequests) {
+        String jimfsSeparator = jimfs.getSeparator();
+        List<String> destinations =
+            actualRequests.stream().map(u -> u.destination().toString())
+                          .collect(Collectors.toList());
+        List<String> expectedPaths =
+            Arrays.stream(keys).map(k -> DIRECTORY_NAME + jimfsSeparator + k.replace(delimiter, jimfsSeparator)).collect(Collectors.toList());
+        assertThat(destinations).isEqualTo(expectedPaths);
     }
 
     private FileDownload newSuccessfulDownload() {
