+++ services-custom/s3-transfer-manager/src/test/java/software/amazon/awssdk/transfer/s3/internal/UploadDirectoryHelperTest.java
@@ -21,28 +21,38 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
 import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystem;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.function.Function;
+import java.util.stream.Collectors;
+import org.assertj.core.util.Sets;
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.ArgumentCaptor;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.services.s3.internal.crt.S3MetaRequestPauseObservable;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.PutObjectResponse;
+import software.amazon.awssdk.testutils.FileUtils;
 import software.amazon.awssdk.transfer.s3.config.TransferRequestOverrideConfiguration;
 import software.amazon.awssdk.transfer.s3.internal.model.DefaultFileUpload;
 import software.amazon.awssdk.transfer.s3.internal.progress.DefaultTransferProgress;
@@ -57,33 +67,52 @@ import software.amazon.awssdk.transfer.s3.progress.LoggingTransferListener;
 import software.amazon.awssdk.transfer.s3.progress.TransferListener;
 
 public class UploadDirectoryHelperTest {
-    private static FileSystem jimfs;
-    private static Path directory;
+    private FileSystem jimfs;
+    private Path directory;
+
+    /**
+     * Local directory is needed to test symlinks because jimfs doesn't work well with symlinks
+     */
+    private static Path localDirectory;
     private Function<UploadFileRequest, FileUpload> singleUploadFunction;
     private UploadDirectoryHelper uploadDirectoryHelper;
 
+    public static Collection<FileSystem> fileSystems() {
+        return Arrays.asList(Jimfs.newFileSystem(Configuration.unix()),
+                             Jimfs.newFileSystem(Configuration.osX()),
+                             Jimfs.newFileSystem(Configuration.windows()));
+    }
+
     @BeforeAll
     public static void setUp() throws IOException {
+        localDirectory = createLocalTestDirectory();
+    }
+
+    @AfterAll
+    public static void tearDown() throws IOException {
+        FileUtils.cleanUpTestDirectory(localDirectory);
+    }
+
+    @BeforeEach
+    public void methodSetup() throws IOException {
         jimfs = Jimfs.newFileSystem();
         directory = jimfs.getPath("test");
         Files.createDirectory(directory);
         Files.createFile(jimfs.getPath("test/1"));
         Files.createFile(jimfs.getPath("test/2"));
-    }
 
-    @AfterAll
-    public static void tearDown() throws IOException {
-        jimfs.close();
-    }
-
-    @BeforeEach
-    public void methodSetup() {
         singleUploadFunction = mock(Function.class);
+
         uploadDirectoryHelper = new UploadDirectoryHelper(TransferManagerConfiguration.builder().build(), singleUploadFunction);
     }
 
+    @AfterEach
+    public void methodCleanup() throws IOException {
+        jimfs.close();
+    }
+
     @Test
-    public void uploadDirectory_cancel_shouldCancelAllFutures() {
+    void uploadDirectory_cancel_shouldCancelAllFutures() {
         CompletableFuture<CompletedFileUpload> future = new CompletableFuture<>();
         FileUpload fileUpload = newUpload(future);
 
@@ -108,7 +137,7 @@ public class UploadDirectoryHelperTest {
     }
 
     @Test
-    public void uploadDirectory_allUploadsSucceed_failedUploadsShouldBeEmpty() throws Exception {
+    void uploadDirectory_allUploadsSucceed_failedUploadsShouldBeEmpty() throws Exception {
         PutObjectResponse putObjectResponse = PutObjectResponse.builder().eTag("1234").build();
         CompletedFileUpload completedFileUpload = CompletedFileUpload.builder().response(putObjectResponse).build();
         CompletableFuture<CompletedFileUpload> successfulFuture = new CompletableFuture<>();
@@ -136,7 +165,7 @@ public class UploadDirectoryHelperTest {
     }
 
     @Test
-    public void uploadDirectory_partialSuccess_shouldProvideFailedUploads() throws Exception {
+    void uploadDirectory_partialSuccess_shouldProvideFailedUploads() throws Exception {
         PutObjectResponse putObjectResponse = PutObjectResponse.builder().eTag("1234").build();
         CompletedFileUpload completedFileUpload = CompletedFileUpload.builder().response(putObjectResponse).build();
         CompletableFuture<CompletedFileUpload> successfulFuture = new CompletableFuture<>();
@@ -165,7 +194,7 @@ public class UploadDirectoryHelperTest {
     }
 
     @Test
-    public void uploadDirectory_withRequestTransformer_usesRequestTransformer() throws Exception {
+    void uploadDirectory_withRequestTransformer_usesRequestTransformer() throws Exception {
         PutObjectResponse putObjectResponse = PutObjectResponse.builder().eTag("1234").build();
         CompletedFileUpload completedFileUpload = CompletedFileUpload.builder().response(putObjectResponse).build();
         CompletableFuture<CompletedFileUpload> successfulFuture = new CompletableFuture<>();
@@ -190,8 +219,8 @@ public class UploadDirectoryHelperTest {
         List<TransferListener> listeners = Arrays.asList(LoggingTransferListener.create());
 
         Consumer<UploadFileRequest.Builder> uploadFileRequestTransformer = r -> r.source(newSource)
-                                                                    .putObjectRequest(newPutObjectRequest)
-                                                                    .transferListeners(listeners);
+                                                                                 .putObjectRequest(newPutObjectRequest)
+                                                                                 .transferListeners(listeners);
 
         uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
                                                                     .sourceDirectory(directory)
@@ -215,6 +244,212 @@ public class UploadDirectoryHelperTest {
         });
     }
 
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void uploadDirectory_defaultSetting_shouldRecursivelyUpload(FileSystem fileSystem) {
+        directory = createJimFsTestDirectory(fileSystem);
+        ArgumentCaptor<UploadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(UploadFileRequest.class);
+
+        when(singleUploadFunction.apply(requestArgumentCaptor.capture()))
+            .thenReturn(completedUpload());
+        DirectoryUpload uploadDirectory =
+            uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                        .sourceDirectory(directory)
+                                                                        .bucket("bucket")
+                                                                        .followSymbolicLinks(false)
+                                                                        .build());
+        uploadDirectory.completionFuture().join();
+
+        List<UploadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+        actualRequests.forEach(r -> assertThat(r.putObjectRequest().bucket()).isEqualTo("bucket"));
+
+        assertThat(actualRequests.size()).isEqualTo(3);
+
+        List<String> keys =
+            actualRequests.stream().map(u -> u.putObjectRequest().key())
+                          .collect(Collectors.toList());
+
+        assertThat(keys).containsOnly("bar.txt", "foo/1.txt", "foo/2.txt");
+    }
+
+    @Test
+    void uploadDirectory_depth1FollowSymlinkTrue_shouldOnlyUploadTopLevel() {
+        ArgumentCaptor<UploadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(UploadFileRequest.class);
+
+        when(singleUploadFunction.apply(requestArgumentCaptor.capture())).thenReturn(completedUpload());
+        DirectoryUpload uploadDirectory =
+            uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                        .sourceDirectory(localDirectory)
+                                                                        .bucket("bucket")
+                                                                        .maxDepth(1)
+                                                                        .followSymbolicLinks(true)
+                                                                        .build());
+        uploadDirectory.completionFuture().join();
+
+        List<UploadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+        List<String> keys =
+            actualRequests.stream().map(u -> u.putObjectRequest().key())
+                          .collect(Collectors.toList());
+
+        assertThat(keys.size()).isEqualTo(2);
+        assertThat(keys).containsOnly("bar.txt", "symlink2");
+    }
+
+    @Test
+    void uploadDirectory_FollowSymlinkTrue_shouldIncludeLinkedFiles() {
+        ArgumentCaptor<UploadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(UploadFileRequest.class);
+
+        when(singleUploadFunction.apply(requestArgumentCaptor.capture())).thenReturn(completedUpload());
+        DirectoryUpload uploadDirectory =
+            uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                        .sourceDirectory(localDirectory)
+                                                                        .bucket("bucket")
+                                                                        .followSymbolicLinks(true)
+                                                                        .build());
+        uploadDirectory.completionFuture().join();
+
+        List<UploadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+        actualRequests.forEach(r -> assertThat(r.putObjectRequest().bucket()).isEqualTo("bucket"));
+
+        List<String> keys =
+            actualRequests.stream().map(u -> u.putObjectRequest().key())
+                          .collect(Collectors.toList());
+
+        assertThat(keys.size()).isEqualTo(5);
+        assertThat(keys).containsOnly("bar.txt", "foo/1.txt", "foo/2.txt", "symlink/2.txt", "symlink2");
+    }
+
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void uploadDirectory_withPrefix_keysShouldHavePrefix(FileSystem fileSystem) {
+        directory = createJimFsTestDirectory(fileSystem);
+        ArgumentCaptor<UploadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(UploadFileRequest.class);
+
+        when(singleUploadFunction.apply(requestArgumentCaptor.capture())).thenReturn(completedUpload());
+        DirectoryUpload uploadDirectory =
+            uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                        .sourceDirectory(directory)
+                                                                        .bucket("bucket")
+                                                                        .s3Prefix("yolo")
+                                                                        .build());
+        uploadDirectory.completionFuture().join();
+
+        List<String> keys =
+            requestArgumentCaptor.getAllValues().stream().map(u -> u.putObjectRequest().key())
+                                 .collect(Collectors.toList());
+
+        assertThat(keys.size()).isEqualTo(3);
+        keys.forEach(r -> assertThat(r).startsWith("yolo/"));
+    }
+
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void uploadDirectory_withDelimiter_shouldHonor(FileSystem fileSystem) {
+        directory = createJimFsTestDirectory(fileSystem);
+        ArgumentCaptor<UploadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(UploadFileRequest.class);
+
+        when(singleUploadFunction.apply(requestArgumentCaptor.capture())).thenReturn(completedUpload());
+        DirectoryUpload uploadDirectory =
+            uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                        .sourceDirectory(directory)
+                                                                        .bucket("bucket")
+                                                                        .s3Delimiter(",")
+                                                                        .s3Prefix("yolo")
+                                                                        .build());
+        uploadDirectory.completionFuture().join();
+
+        List<String> keys =
+            requestArgumentCaptor.getAllValues().stream().map(u -> u.putObjectRequest().key())
+                                 .collect(Collectors.toList());
+
+        assertThat(keys.size()).isEqualTo(3);
+        assertThat(keys).containsOnly("yolo,foo,2.txt", "yolo,foo,1.txt", "yolo,bar.txt");
+    }
+
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void uploadDirectory_maxLengthOne_shouldOnlyUploadTopLevel(FileSystem fileSystem) {
+        directory = createJimFsTestDirectory(fileSystem);
+        ArgumentCaptor<UploadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(UploadFileRequest.class);
+
+        when(singleUploadFunction.apply(requestArgumentCaptor.capture()))
+            .thenReturn(completedUpload());
+        DirectoryUpload uploadDirectory =
+            uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                        .sourceDirectory(directory)
+                                                                        .bucket("bucket")
+                                                                        .maxDepth(1)
+                                                                        .build());
+        uploadDirectory.completionFuture().join();
+
+        List<UploadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+        actualRequests.forEach(r -> assertThat(r.putObjectRequest().bucket()).isEqualTo("bucket"));
+
+        assertThat(actualRequests.size()).isEqualTo(1);
+
+        List<String> keys =
+            actualRequests.stream().map(u -> u.putObjectRequest().key())
+                          .collect(Collectors.toList());
+
+        assertThat(keys).containsOnly("bar.txt");
+    }
+
+
+    @ParameterizedTest
+    @MethodSource("fileSystems")
+    void uploadDirectory_directoryNotExist_shouldCompleteFutureExceptionally(FileSystem fileSystem) {
+        directory = createJimFsTestDirectory(fileSystem);
+        assertThatThrownBy(() -> uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder().sourceDirectory(Paths.get(
+                                                                                                 "randomstringneverexistas234ersaf1231"))
+                                                                                             .bucket("bucketName").build()).completionFuture().join())
+            .hasMessageContaining("does not exist").hasCauseInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void uploadDirectory_notDirectory_shouldCompleteFutureExceptionally() {
+        assertThatThrownBy(() -> uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                                             .sourceDirectory(Paths.get(localDirectory.toString(), "symlink"))
+                                                                                             .bucket("bucketName").build()).completionFuture().join())
+            .hasMessageContaining("is not a directory").hasCauseInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void uploadDirectory_notDirectoryFollowSymlinkTrue_shouldCompleteSuccessfully() {
+        ArgumentCaptor<UploadFileRequest> requestArgumentCaptor = ArgumentCaptor.forClass(UploadFileRequest.class);
+
+        when(singleUploadFunction.apply(requestArgumentCaptor.capture())).thenReturn(completedUpload());
+        DirectoryUpload uploadDirectory = uploadDirectoryHelper.uploadDirectory(UploadDirectoryRequest.builder()
+                                                                                                      .followSymbolicLinks(true)
+                                                                                                      .sourceDirectory(Paths.get(localDirectory.toString(), "symlink"))
+                                                                                                      .bucket("bucket").build());
+
+        uploadDirectory.completionFuture().join();
+
+        List<UploadFileRequest> actualRequests = requestArgumentCaptor.getAllValues();
+        actualRequests.forEach(r -> assertThat(r.putObjectRequest().bucket()).isEqualTo("bucket"));
+
+        assertThat(actualRequests.size()).isEqualTo(1);
+
+        List<String> keys =
+            actualRequests.stream().map(u -> u.putObjectRequest().key())
+                          .collect(Collectors.toList());
+
+        assertThat(keys).containsOnly("2.txt");
+    }
+
+    private DefaultFileUpload completedUpload() {
+        return new DefaultFileUpload(CompletableFuture.completedFuture(CompletedFileUpload.builder()
+                                                                                          .response(PutObjectResponse.builder().build())
+                                                                                          .build()),
+                                     new DefaultTransferProgress(DefaultTransferProgressSnapshot.builder()
+                                                                                                .transferredBytes(0L)
+                                                                                                .build()),
+                                     new S3MetaRequestPauseObservable(),
+                                     UploadFileRequest.builder()
+                                                      .source(Paths.get(".")).putObjectRequest(b -> b.bucket("bucket").key("key"))
+                                                      .build());
+    }
+
     private FileUpload newUpload(CompletableFuture<CompletedFileUpload> future) {
         return new DefaultFileUpload(future,
                                      new DefaultTransferProgress(DefaultTransferProgressSnapshot.builder()
@@ -222,7 +457,70 @@ public class UploadDirectoryHelperTest {
                                                                                                 .build()),
                                      new S3MetaRequestPauseObservable(),
                                      UploadFileRequest.builder()
-                                                      .putObjectRequest(p -> p.key("key").bucket("bucket")).source(Paths.get("test.txt"))
+                                                      .putObjectRequest(p -> p.key("key").bucket("bucket")).source(Paths.get(
+                                                          "test.txt"))
                                                       .build());
+    }
+
+    private Path createJimFsTestDirectory(FileSystem fileSystem) {
+
+        try {
+            return createJmfsTestDirectory(fileSystem);
+        } catch (IOException exception) {
+            throw new UncheckedIOException(exception);
+        }
+
+    }
+
+    private static Path createLocalTestDirectory() {
+
+        try {
+            return createLocalTestDirectoryWithSymLink();
+        } catch (IOException exception) {
+            throw new UncheckedIOException(exception);
+        }
+    }
+
+    /**
+     * Create a test directory with the following structure - test1 - foo - 1.txt - 2.txt - bar.txt - symlink -> test2 - symlink2
+     * -> test3/4.txt - test2 - 2.txt - test3 - 4.txt
+     */
+    private static Path createLocalTestDirectoryWithSymLink() throws IOException {
+        Path directory = Files.createTempDirectory("test1");
+        Path anotherDirectory = Files.createTempDirectory("test2");
+        Path thirdDirectory = Files.createTempDirectory("test3");
+
+        String directoryName = directory.toString();
+        String anotherDirectoryName = anotherDirectory.toString();
+
+        Files.createDirectory(Paths.get(directory + "/foo"));
+
+        Files.write(Paths.get(directoryName, "bar.txt"), "bar".getBytes(StandardCharsets.UTF_8));
+        Files.write(Paths.get(directoryName, "foo/1.txt"), "1".getBytes(StandardCharsets.UTF_8));
+        Files.write(Paths.get(directoryName, "foo/2.txt"), "2".getBytes(StandardCharsets.UTF_8));
+
+        Files.write(Paths.get(anotherDirectoryName, "2.txt"), "2".getBytes(StandardCharsets.UTF_8));
+        Files.write(Paths.get(thirdDirectory.toString(), "3.txt"), "3".getBytes(StandardCharsets.UTF_8));
+
+        Files.createSymbolicLink(Paths.get(directoryName, "symlink"), anotherDirectory);
+        Files.createSymbolicLink(Paths.get(directoryName, "symlink2"), Paths.get(thirdDirectory.toString(), "3.txt"));
+        return directory;
+    }
+
+    /**
+     * Create a test directory with the following structure - test1 - foo - 1.txt - 2.txt - bar.txt
+     */
+    private Path createJmfsTestDirectory(FileSystem jimfs) throws IOException {
+        String directoryName = "test";
+        Path directory = jimfs.getPath(directoryName);
+
+        Files.createDirectory(directory);
+
+        Files.createDirectory(jimfs.getPath(directoryName + "/foo"));
+
+        Files.write(jimfs.getPath(directoryName, "bar.txt"), "bar".getBytes(StandardCharsets.UTF_8));
+        Files.write(jimfs.getPath(directoryName, "foo", "1.txt"), "1".getBytes(StandardCharsets.UTF_8));
+        Files.write(jimfs.getPath(directoryName, "foo", "2.txt"), "2".getBytes(StandardCharsets.UTF_8));
+        return directory;
     }
 }
