+++ services-custom/s3-transfer-manager/src/test/resources/log4j2.properties
@@ -28,8 +28,8 @@ rootLogger.appenderRef.stdout.ref = ConsoleAppender
 #logger.sdk.name = software.amazon.awssdk
 #logger.sdk.level = debug
 #
-logger.tm.name = software.amazon.awssdk.transfer.s3
-logger.tm.level = debug
+#logger.tm.name = software.amazon.awssdk.transfer.s3
+#logger.tm.level = info
 #
 #logger.request.name = software.amazon.awssdk.request
 #logger.request.level = debug
